requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow
opencv-python
streamlit
pillow
tensor
tensorflow